services/database-migration/pyproject.toml
[project]
name = "database-migration"
version = "0.1.0"
description = "Database migration service for FleetManager"
requires-python = ">=3.13"
dependencies = [
    "database-models",
    "alembic>=1.13.0",
    "opentelemetry-api>=1.25.0",
    "opentelemetry-sdk>=1.25.0",
    "opentelemetry-exporter-otlp>=1.25.0",
]

[tool.uv.sources]
database-models = { workspace = true }
